Integration of new opencl stuff

+++ Camera/POVRaySimulatedCameraDevice.py
@@ -195,7 +195,7 @@ class POVRaySimulatedCameraDevice:
         #print a
         a = a.mean(2)
         
-        self.im_array = a.astype(float)
+        self.im_array = a.astype(float32)
         #self.im_array = self.im_array[:, 1:]
                 
         # start the analysis process
